# docs/swagger.yaml
definitions:
  controllers.ForumResponse:
    properties:
      forums:
        items:
          $ref: '#/definitions/models.Forum'
        type: array
      user_detail:
        $ref: '#/definitions/models.UserDetail'
    type: object
  controllers.LoginResponse:
    properties:
      token:
        type: string
      user_id:
        type: string
    type: object
...
  models.Forum:
    properties:
      admin_list:
        items:
          type: integer
        type: array
      cover:
        type: string
      create_at:
        type: string
      description:
        type: string
      forum_id:
        type: integer
      forum_name:
        type: string
      is_public:
        type: boolean
      post_num:
        type: integer
      subscribe_num:
        type: integer
    type: object
...
  models.User:
    properties:
      avatar:
        type: string
      create_at:
        type: string
      email:
        type: string
      is_admin:
        type: boolean
      password:
        type: string
      user_id:
        type: integer
      username:
        type: string
    type: object
  models.UserDetail:
    properties:
      avatar:
        type: string
      create_at:
        type: string
      email:
        type: string
      is_admin:
        type: boolean
      like_list:
        items:
          type: integer
        type: array
      participate_list:
        items:
          type: integer
        type: array
      password:
        type: string
      star_list:
        items:
          type: integer
        type: array
      user_id:
        type: integer
      username:
        type: string
    type: object
  responses.StatusBadRequestResponse:
    properties:
      code:
        example: 400
        type: integer
      data:
        type: object
      msg:
        type: string
    type: object
  responses.StatusForbiddenResponse:
    properties:
      code:
        example: 403
        type: integer
      data:
        type: object
      msg:
        type: string
    type: object
  responses.StatusInternalServerError:
    properties:
      code:
        example: 500
        type: integer
      data:
        type: object
      msg:
        type: string
    type: object
  responses.StatusOKResponse:
    properties:
      code:
        example: 200
        type: integer
      data:
        type: object
      msg:
        type: string
    type: object
info:
  contact: {}
paths:
  /forums:
    get:
      consumes:
      - application/json
      description: GetAllPublicFroums
      produces:
      - application/json
      responses:
        "200":
          description: 获取全部公开论坛
          schema:
            allOf:
            - $ref: '#/definitions/responses.StatusOKResponse'
            - properties:
                data:
                  $ref: '#/definitions/controllers.ForumResponse'
              type: object
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/responses.StatusInternalServerError'
      summary: GetAllPublicFroums
      tags:
      - Forums
    post:
      consumes:
      - application/json
      description: CreateForum
      parameters:
      - description: 论坛名
        in: body
        name: forum_name
        required: true
        schema:
          type: string
      - description: 是否公开
        in: body
        name: is_public
        required: true
        schema:
          type: boolean
      - description: 论坛描述
        in: body
        name: description
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: 论坛创建成功
          schema:
            $ref: '#/definitions/responses.StatusOKResponse'
        "400":
          description: 参数不合法
          schema:
            $ref: '#/definitions/responses.StatusBadRequestResponse'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/responses.StatusInternalServerError'
      summary: CreateForum
      tags:
      - Forums
  /forums/{forum_id}:
    get:
      consumes:
      - application/json
      description: GetForumByID
      produces:
      - application/json
      responses:
        "200":
          description: 获取全部公开论坛
          schema:
            allOf:
            - $ref: '#/definitions/responses.StatusOKResponse'
            - properties:
                data:
                  $ref: '#/definitions/controllers.ForumResponse'
              type: object
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/responses.StatusInternalServerError'
      summary: GetForumByID
      tags:
      - Forums
...
  /users:
    get:
      consumes:
      - application/json
      description: GetAllUsers
      parameters:
      - description: 将token放在请求头部的‘Authorization‘字段中，并以‘Bearer ‘开头
        in: header
        name: token
        required: true
        type: string
      - description: 用户名的子串
        in: header
        name: username
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取全部用户
          schema:
            allOf:
            - $ref: '#/definitions/responses.StatusOKResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.User'
                  type: array
              type: object
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/responses.StatusInternalServerError'
      summary: GetAllUsers
      tags:
      - Users
    post:
      consumes:
      - application/json
      description: UserRegister
      parameters:
      - description: 用户名
        in: body
        name: username
        required: true
        schema:
          type: string
      - description: 密码
        in: body
        name: password
        required: true
        schema:
          type: string
      - description: 邮箱
        in: body
        name: email
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: 注册成功
          schema:
            $ref: '#/definitions/responses.StatusOKResponse'
        "400":
          description: 参数不合法
          schema:
            $ref: '#/definitions/responses.StatusBadRequestResponse'
        "403":
          description: 该邮箱已经被使用
          schema:
            $ref: '#/definitions/responses.StatusForbiddenResponse'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/responses.StatusInternalServerError'
      summary: UserRegister
      tags:
      - Users
    put:
      consumes:
      - application/json
      description: UserLogin
      parameters:
      - description: 用户名或者邮箱
        in: body
        name: input
        required: true
        schema:
          type: string
      - description: 密码
        in: body
        name: password
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: 正确登陆
          schema:
            allOf:
            - $ref: '#/definitions/responses.StatusOKResponse'
            - properties:
                data:
                  $ref: '#/definitions/controllers.LoginResponse'
              type: object
        "400":
          description: 参数不合法
          schema:
            $ref: '#/definitions/responses.StatusBadRequestResponse'
        "403":
          description: 该用户名或邮箱不存在
          schema:
            $ref: '#/definitions/responses.StatusForbiddenResponse'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/responses.StatusInternalServerError'
      summary: UserLogin
      tags:
      - Users
...
swagger: "2.0"